feat: skadi api client v2
