+++ __tests__/campaigns.ts
@@ -749,7 +749,7 @@ describe('mutation startCampaign', () => {
         { flags: updateFlagsStatement<Post>({ campaignId: null }) },
       );
 
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/create', (body) => {
         const uuidRegex =
           /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -788,7 +788,7 @@ describe('mutation startCampaign', () => {
 
   it('should handle transfer failure gracefully', async () => {
     loggedUser = '1';
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/create', (body) => {
         const uuidRegex =
           /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -894,7 +894,7 @@ describe('mutation startCampaign', () => {
 
     it('should successfully start post campaign', async () => {
       loggedUser = '1';
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/create', (body) => {
           const uuidRegex =
             /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -977,7 +977,7 @@ describe('mutation startCampaign', () => {
 
     it('should successfully start source campaign', async () => {
       loggedUser = '3'; // moderator in 'm'
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/create', (body) => {
           const uuidRegex =
             /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -1045,7 +1045,7 @@ describe('mutation startCampaign', () => {
         { postId: 'mp1', keyword: 'squadtag4' },
       ]);
 
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/create', (body) => {
           const uuidRegex =
             /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -1114,7 +1114,7 @@ describe('mutation startCampaign', () => {
         { postId: 'mp2', keyword: 'three' },
       ]);
 
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/create', (body) => {
           const uuidRegex =
             /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
@@ -1217,7 +1217,7 @@ describe('mutation stopCampaign', () => {
         { flags: updateFlagsStatement<Post>({ campaignId: CAMPAIGN_UUID_1 }) },
       );
 
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/cancel', (body) => {
         return (
           body?.campaign_id === CAMPAIGN_UUID_1 &&
@@ -1274,7 +1274,7 @@ describe('mutation stopCampaign', () => {
       },
     );
 
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/cancel', (body) => {
         return (
           body?.campaign_id === CAMPAIGN_UUID_3 &&
@@ -1328,7 +1328,7 @@ describe('mutation stopCampaign', () => {
         { flags: updateFlagsStatement<Post>({ campaignId: CAMPAIGN_UUID_2 }) },
       );
 
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/cancel', (body) => {
         return (
           body?.campaign_id === CAMPAIGN_UUID_2 &&
@@ -1374,7 +1374,7 @@ describe('mutation stopCampaign', () => {
         { flags: updateFlagsStatement<Post>({ campaignId: CAMPAIGN_UUID_5 }) },
       );
 
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/cancel', (body) => {
         return (
           body?.campaign_id === CAMPAIGN_UUID_5 &&
@@ -1588,7 +1588,7 @@ describe('query dailyCampaignReachEstimate', () => {
 
     it('should accept valid budget values and make correct HTTP call', async () => {
       // Mock the HTTP response using nock
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/reach', (body) => {
           const keywords = body?.targeting?.value?.boost?.keywords || [];
           return (
@@ -1621,7 +1621,7 @@ describe('query dailyCampaignReachEstimate', () => {
 
     it('should handle minimum budget value (1000 cores)', async () => {
       // Mock the HTTP response using nock
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/reach', (body) => {
           const keywords = body?.targeting?.value?.boost?.keywords || [];
           return (
@@ -1654,7 +1654,7 @@ describe('query dailyCampaignReachEstimate', () => {
 
     it('should handle maximum budget value (100000 cores)', async () => {
       // Mock the HTTP response using nock
-      nock(process.env.SKADI_API_ORIGIN)
+      nock(process.env.SKADI_API_ORIGIN_V2)
         .post('/campaign/reach', (body) => {
           const keywords = body?.targeting?.value?.boost?.keywords || [];
           return (
@@ -1690,7 +1690,7 @@ describe('query dailyCampaignReachEstimate', () => {
     loggedUser = '1';
 
     // Mock the HTTP response using nock
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         const keywords = body?.targeting?.value?.boost?.keywords || [];
         return (
@@ -1734,7 +1734,7 @@ describe('query dailyCampaignReachEstimate', () => {
     );
 
     // Mock the HTTP response using nock
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         const keywords = body?.targeting?.value?.boost?.keywords || [];
         return (
@@ -1769,7 +1769,7 @@ describe('query dailyCampaignReachEstimate', () => {
     loggedUser = '3'; // moderator in 'm'
 
     // Mock the HTTP response using nock
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         const keywords = body?.targeting?.value?.boost?.keywords || [];
         return (
@@ -1803,7 +1803,7 @@ describe('query dailyCampaignReachEstimate', () => {
     loggedUser = '1';
 
     // Mock the HTTP response where min and max impressions are the same
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         const keywords = body?.targeting?.value?.boost?.keywords || [];
         return (
@@ -1839,7 +1839,7 @@ describe('query dailyCampaignReachEstimate', () => {
   it('should work correctly for both post and source campaign types', async () => {
     // Test post campaign
     loggedUser = '1';
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         return (
           body.budget === 30 &&
@@ -1867,7 +1867,7 @@ describe('query dailyCampaignReachEstimate', () => {
 
     // Test source campaign
     loggedUser = '3';
-    nock(process.env.SKADI_API_ORIGIN)
+    nock(process.env.SKADI_API_ORIGIN_V2)
       .post('/campaign/reach', (body) => {
         const keywords = body?.targeting?.value?.boost?.keywords || [];
         return (
