+++ src/common/campaign/source.ts
@@ -28,9 +28,9 @@ import {
   UserTransactionStatus,
   UserTransactionType,
 } from '../../entity/user/UserTransaction';
-import { skadiApiClientV1 } from '../../integrations/skadi/api/v1/clients';
 import { usdToCores } from '../number';
 import { systemUser, updateFlagsStatement } from '../utils';
+import { skadiApiClientV2 } from '../../integrations/skadi/api/v2/clients';
 
 export const validateSquadBoostPermissions = async (
   ctx: AuthContext,
@@ -112,7 +112,7 @@ export const startCampaignSource = async (props: StartCampaignMutationArgs) => {
     const last30tags = await getSourceTags(manager, source.id);
     const finalTags = last30tags.length > 3 ? last30tags : []; // when it is 3 or below, we set global targeting
 
-    await skadiApiClientV1.startCampaign(campaign, finalTags);
+    await skadiApiClientV2.startCampaign(campaign, finalTags);
 
     await manager
       .getRepository(Source)
@@ -156,7 +156,7 @@ export const stopCampaignSource = async ({
 }: StopCampaignProps) => {
   const { id: campaignId, userId, referenceId } = campaign;
 
-  const { budget } = await skadiApiClientV1.cancelCampaign({
+  const { budget } = await skadiApiClientV2.cancelCampaign({
     campaignId,
     userId,
   });
