+++ src/integrations/skadi/api/v2/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export { SkadiApiClientV2 } from './clients';
